Improved the search UI.
Made sure the date navigation bar disappears when the user activates search. Fixes #37.

Search in the schedule view will now search across all dates. Fixes #43.

As per Heiko's suggestion, the search bar will be activated immediately when the user taps on the search button. However, I'm not sure this is a cool idea - it looks a bit odd. Fixes #44.

The search bar will be hidden initially. Fixes #21.

The date navigator is now a separate class, allowing better re-use and also making for a cleaner architecture.